Config: Cut AppConfig into Config and AppConfig
so that the app can have its own configuration
