+++ SiriIntents/IntentHandler.m
@@ -36,7 +36,7 @@
     if (self)
     {
         // Set static application settings
-        [AppConfig.shared setupSettings];
+        [Config.shared setupSettings];
 
         // NSLog -> console.log file when not debugging the app
         if (!isatty(STDERR_FILENO))
